1.0.1 -> 1.0.2 整数値をLong型として取り扱うよう変更。

+++ src/main/groovy/com/sorabito/gexcelparser/GExcelParser.groovy
@@ -165,11 +165,11 @@ class GExcelParser {
             return cell.getDateCellValue()
         }
 
-        // double or int
+        // double or long
         DataFormatter formatter = new DataFormatter()
         String retValue = formatter.formatCellValue(cell)
 
         if (retValue.contains('.')) { return Double.parseDouble(retValue) }
-        return Integer.parseInt(retValue)
+        return Long.parseLong(retValue)
     }
 }
